+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 2024-03-28
 
 - 实现用户认证算法 | Implement user authentication algorithm
+- 实现初始化数据库表 `Staff` 的功能 | Implement the function of initializing the database table `Staff`
 
 ## 2024-03-27